src/main/java/com/epam/microservice/security/AuthTokenFilter.java
package com.epam.microservice.security;

import jakarta.servlet.FilterChain;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import lombok.AllArgsConstructor;
import org.springframework.web.filter.OncePerRequestFilter;
import java.io.IOException;

@AllArgsConstructor
public class AuthTokenFilter extends OncePerRequestFilter {
    private final JwtService jwtService;

    @Override
    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
            throws ServletException, IOException {
        String jwt = jwtService.parseJwt(request);
        if (jwt == null || !jwtService.validateJwtToken(jwt)) {
            response.sendError(HttpServletResponse.SC_FORBIDDEN, "Invalid token");
        }
        filterChain.doFilter(request, response);
    }
}
